Update readme-demo dependencies
This change brings some dependencies of readme-demo up to date.

It requires Node 18, which is the current long term support version.

It requires the current versions of @fastly/js-compute and webpack.

It removes the unnecessary dependency of core-js.

It uses npm and npm scripts.

It silences warning from webpack.

It updates the documentation to match the above.